Break apart mh_data.h and data.c into multiple header/source files so that not all structs are imported when only a subset is needed + so we can separate some of those structs from the host-only fourier type (needed for cuFFT interface).
