--- include/mh_parameters.h
@@ -1,10 +1,0 @@
-/* mugy: mh_parameters.h
-
-   Global parameters in mugy.
-*/
-
-#include "mh_data.h"
-
-extern mint nFrames, mpiProcs[nDim];
-extern real kxMaxDyn[nDim], dt, endTime;
-
