+++ src/data.c
@@ -1,4 +1,4 @@
-/* mugy: data_mugy
+/* mugy: data.c
    
    Parameters and fields used throughout mugy.
 */
@@ -6,80 +6,6 @@
 #include "mh_data.h"
 #include "mh_data_dev.h"
 #include <string.h>
-
-real* getMoment_real(struct mugy_realGrid grid, struct mugy_population pop, mint sIdx, mint momIdx, real *momIn) {
-  // Return a pointer to the momIdx-th moment of the sIdx-th species in mom.
-  real* ptrOut = momIn;
-  mint momOff = 0;
-  for (mint s=0; s<sIdx; s++) momOff += pop.spec[s].numMoments;
-  return ptrOut+(momOff+momIdx)*grid.NxTot;
-}
-
-fourier* getMoment_fourier(struct mugy_fourierGrid grid, struct mugy_population pop, mint sIdx, mint momIdx, fourier *momkIn) {
-  // Return a pointer to the momIdx-th moment of the sIdx-th species in momk.
-  fourier* ptrOut = momkIn;
-  mint momOff = 0;
-  for (mint s=0; s<sIdx; s++) momOff += pop.spec[s].numMoments;
-  return ptrOut+(momOff+momIdx)*grid.NekxTot;
-}
-
-mint sub2lin_real(mint *xI, const struct mugy_realGrid grid) {
-  // Given the nDim-dimensional index (subscript) xI return the linear index
-  // in a real grid. We assume row major order for the (z,x,y) dimensions.
-  mint strides[nDim] = {grid.Nx[1],1,grid.NxyTot};
-  mint lin;
-  for (mint d=0; d<nDim; d++) lin += xI[d]*strides[d];
-  return lin;
-}
-
-mint sub2lin_fourier(mint *kxI, const struct mugy_fourierGrid grid) {
-  // Given the nDim-dimensional index (subscript) kxI return the linear index
-  // in a Fourier grid. We assume row major order for the (kz,kx,ky) dimensions.
-  mint strides[nDim] = {grid.Nekx[1],1,grid.NekxyTot};
-  mint lin;
-  for (mint d=0; d<nDim; d++) lin += kxI[d]*strides[d];
-  return lin;
-}
-
-void lin2sub_real(mint *xI, mint lin, const struct mugy_realGrid grid) {
-  // Given the linear index 'lin' in a real grid, return the nDim-dimensional
-  // index (subscript) xI. We assume row major order for the (z,x,y) dimensions.
-  mint strides[nDim] = {grid.Nx[1],1,grid.NxyTot};
-  for (mint d=0; d<nDim; d++) {
-    xI[d] = lin/strides[d];
-    lin -= xI[d]*strides[d];
-  }
-}
-
-void lin2sub_fourier(mint *kxI, mint lin, const struct mugy_fourierGrid grid) {
-  // Given the linear index 'lin' in a Fourier grid, return the nDim-dimensional
-  // index (subscript) kxI. We assume row major order for the (kz,kx,ky) dimensions.
-  mint strides[nDim] = {grid.Nekx[1],1,grid.NekxyTot};
-  for (mint d=0; d<nDim; d++) {
-    kxI[d] = lin/strides[d];
-    lin -= kxI[d]*strides[d];
-  }
-}
-
-void get_x(real *x, mint *xI, const struct mugy_realGrid grid) {
-  // Obtain the x=(x,y,z) coordinates given the multidimensional
-  // xI index. Assume the flat grid.x array is organized as {x,y,z}.
-  real* x_p = grid.x;
-  for (mint d=0; d<nDim; d++) {
-    x[d] = x_p[xI[d]]; 
-    x_p += grid.Nx[d];
-  }
-}
-
-void get_kx(real *kx, mint *kxI, const struct mugy_fourierGrid grid) {
-  // Obtain the kx=(kx,ky,kz) coordinates given the multidimensional
-  // kxI index. Assume the flat grid.kx array is organized as {kx,ky,kz}.
-  real* kx_p = grid.kx;
-  for (mint d=0; d<nDim; d++) {
-    kx[d] = kx_p[kxI[d]]; 
-    kx_p += grid.Nekx[d];
-  }
-}
 
 // Functions that copy memory on the host.
 void memcpy_mint_ho(mint *dest, mint *src, mint numElements) {
